change the types from int to long

+++ main.c
@@ -7,23 +7,23 @@
 
 
 // ===========================> Functions Prototype <===============================
-void conv1d(const int *data, const int *filter, int *map_out,
+void conv1d(const long *data, const long *filter, long *map_out,
             int input_len, int input_depth, int n_filter);
 
-void deconv1d(int *data, const int *filter, int *map_out,
+void deconv1d(long *data, const long *filter, long *map_out,
               int input_len, int input_depth, int n_filter);
 
-void skip_add(int *data, int *filter, int *map_out,
+void skip_add(long *data, long *filter, long *map_out,
               int input_len, int input_depth);
 
-void forward_propagation(int *data);
+void forward_propagation(long *data);
 // =================================================================================
 
 int main() {
     printf("EpilepsyGAN!\n");
 
     // allocate memory in CPU for calculation
-    int *eeg_input;
+    long *eeg_input;
     const int in_len = 2048;
     // fill a, b matrices with random values between -16.0f and 16.0f
 //    srand(0);
@@ -35,9 +35,9 @@ int main() {
 }
 
 
-void conv1d(const int *data, const int *filter, int *map_out,
+void conv1d(const long *data, const long *filter, long *map_out,
             const int input_len, const int input_depth, const int n_filter) {
-    int sum;
+    long sum;
     for (int w_n = 0; w_n < n_filter; w_n++) {
         for (int start_index = 1; start_index < input_len; start_index += 2) {
             sum = 0;
@@ -64,7 +64,7 @@ void conv1d(const int *data, const int *filter, int *map_out,
     }
 }
 
-void deconv1d(int *data, const int *filter, int *map_out,
+void deconv1d(long *data, const long *filter, long *map_out,
               int input_len, const int input_depth, const int n_filter) {
     // Upsampling
     input_len *= 2; // Update input len to the upsampled one
@@ -80,7 +80,7 @@ void deconv1d(int *data, const int *filter, int *map_out,
 //    free(data);
 
     // Convolution
-    int sum;
+    long sum;
     for (int w_n = 0; w_n < n_filter; w_n++) {
         for (int start_index = 0; start_index < input_len; start_index++) {
             sum = 0;
@@ -106,7 +106,7 @@ void deconv1d(int *data, const int *filter, int *map_out,
 }
 
 
-void concatenate(const int *data, const int *z, int *map_out,
+void concatenate(const long *data, const long *z, long *map_out,
                  int input_len, int input_depth) {
     for (int w_j = 0; w_j < input_depth; w_j++) {
         for (int w_i = 0; w_i < input_len; w_i++) {
@@ -118,7 +118,7 @@ void concatenate(const int *data, const int *z, int *map_out,
 }
 
 
-void skip_add(int *data, int *filter, int *map_out,
+void skip_add(long *data, long *filter, long *map_out,
               int input_len, const int input_depth) {
     for (int w_j = 0; w_j < input_depth; w_j++) {
         for (int w_i = 0; w_i < input_len; w_i++) {
@@ -126,7 +126,7 @@ void skip_add(int *data, int *filter, int *map_out,
 //            printf("enc: %f\n", mem2d(data, input_len, w_j, w_i));
 //            printf("A: %f\n", mem2d(filter, input_len, w_j, w_i));
 //            printf("dec: %f\n", mem2d(map_out, input_len, w_j, w_i));
-            int add = MUL(mem2d(data, input_len, w_j, w_i), mem2d(filter, input_len, w_j, w_i));
+            long add = MUL(mem2d(data, input_len, w_j, w_i), mem2d(filter, input_len, w_j, w_i));
             if (add <0)
                 add = MUL(add,  INV_LEAKY_RATIO);
             mem2d(map_out, input_len, w_j, w_i) += add;
@@ -138,30 +138,30 @@ void skip_add(int *data, int *filter, int *map_out,
 }
 
 
-void save_file(int *data, char *filename, int input_len) {
+void save_file(long *data, char *filename, int input_len) {
     FILE *fp;
     fp = fopen(filename, "w");
     for (int wi = 0; wi < input_len; wi++)
-        fprintf(fp, "%d\n", data[wi]);
+        fprintf(fp, "%ld\n", data[wi]);
 
     fclose(fp);
 }
 
-void forward_propagation(int *data) {
-    int *encoder_layers_out[8] = {0};
+void forward_propagation(long *data) {
+    long *encoder_layers_out[8] = {0};
     int depth_size[9] = {1, 64, 64, 128, 128, 256, 256, 512, 1024};
     int map_size[9] = {2048, 1024, 512, 256, 128, 64, 32, 16, 8};
 //    float enc_ratio[8] = {1.2602388f, 0.2659299f, 0.1959376f, 0.13729893f, 0.1064626f, 0.0913380f, 0.08377741f, 0.05075028f};
 //    float dec_ratio[8] = {0.06047365f, 0.06849701f, 0.19342509f, 0.3388827f, 0.4225882f, 0.4932240f, 0.5388081f, 2.3255081f};
 
-    int *filter;
-    int *layer_in = (int *) data;
+    long *filter;
+    long *layer_in = (long *) data;
     save_file(layer_in, "input.txt", map_size[0]);
 
     // Encoder
     for (int layer = 0; layer <= 7; layer++) {
         printf("Encoder Layer %d\n", layer);
-        encoder_layers_out[layer] = (int *) malloc(map_size[layer + 1] * depth_size[layer + 1] * sizeof(int));
+        encoder_layers_out[layer] = (long *) malloc(map_size[layer + 1] * depth_size[layer + 1] * sizeof(long));
         filter = enc_w[layer];
         conv1d(layer_in, filter, encoder_layers_out[layer], map_size[layer], depth_size[layer],
                depth_size[layer + 1]);
@@ -173,16 +173,16 @@ void forward_propagation(int *data) {
     }
 
     // Concatenation
-    layer_in = (int *) malloc(map_size[8] * depth_size[8] * 2 * sizeof(int));
+    layer_in = (long *) malloc(map_size[8] * depth_size[8] * 2 * sizeof(long));
     concatenate(encoder_layers_out[7], Z_array, layer_in, map_size[8], depth_size[8]);
     save_file(layer_in, "fxd_concatenate.txt", map_size[8] * depth_size[8] * 2 );
 
     // Decoder
-    int *decoder_layers_out;
-    int *skip;
+    long *decoder_layers_out;
+    long *skip;
     for (int layer = 7; layer >= 0; layer--) {
         printf("Decoder Layer %d\n", layer);
-        decoder_layers_out = (int *) malloc(map_size[layer] * depth_size[layer] * sizeof(int));
+        decoder_layers_out = (long *) malloc(map_size[layer] * depth_size[layer] * sizeof(long));
         filter = dec_w[7 - layer];
         deconv1d(layer_in, filter, decoder_layers_out, map_size[layer + 1],
                  depth_size[layer + 1] * (layer == 7 ? 2 :1), depth_size[layer]);
